add qopt package, introduce IR
Instead of ast->bytecode, do ast->ir->bytecode to simplify
the optimization pass.

The IR is close enough to the bytecode so it's trivial
to assemble it into a bytecode later.

The new qopt package implements IR level optimizations.
Right now it only can remove the redundant const->tmp->arg
and const->tmp->local moves.

Testing is now performed on both optimized and unoptimized
bytecode outputs.

It's still possible to do the fast no-optimization compilation
if compiler is invoked with such options.

For the new CallNative2 benchmark, the introduced optimizations
yield ~10% performance improvement and reduce the func frame size.