add OpZero, re-purpose ReturnFalse and ReturnTrue
Instead of using ReturnFalse and ReturnTrue as bool-only
ops, rename them to ReturnZero and ReturnOne and use them
for int-valued types as well (they have identical repr anyway).

So, instead of this:

	LoadScalarConst tmp0 = 0
	ReturnScalar tmp0

We could generate this:

	ReturnZero

In addition to that, new OpZero can be used to be a shorter
encoding of LoadScalarConst of 0. This is useful for
initialization of zero scalar types as well as any other
zero values that fit into a VM slot (24 bytes).