internal/qruntime: use generic addptr instead of add+addb
Generic function is inferior in Go 1.19 due to the lack
of inlining, but that can be fixed by using GOEXPERIMENT=unified.
Starting with Go 1.20 this is no longer necessary as it becomes a default.

On Go 1.20, this change affects the eval() function code size positively.

Bump CI Go version to 1.18 to allow generics.